usando login required na app cliente

+++ cliente/views.py
@@ -3,22 +3,30 @@ from django.views.generic import ListView
 
 from cliente.models import Cliente
 from cliente.forms import ClienteForm
+from django.contrib.auth.decorators import login_required
+    
+
 
 
 class ListaCliente(ListView):
     model = Cliente
     template_name = 'cliente/pagina-inicial-cliente.html'
 
+    def get_queryset(self):
+        return Cliente.objects.filter(usuario=self.request.user)
 
+
+
+@login_required
 def cadastrarcliente(request):
     template_name = 'cliente/formularios/formulario-cadastrar-cliente.html'
-    form = ClienteForm(request.POST or None)
-
-    if request.method == 'POST':
+    
+    form = ClienteForm(request.POST or None, initial={'usuario': request.user})
+   
+    if request.method == 'POST':  
         if form.is_valid():
             cliente = form.save()
             template_name = 'cliente/tabela/linhas-tabela-cliente.html'
-
             context = {'object': cliente}
             return render(request, template_name, context)
 
@@ -26,10 +34,15 @@ def cadastrarcliente(request):
     return render(request, template_name, context)
 
 
+
+@login_required
 def editarcliente(request, pk):
     template_name = 'cliente/formularios/formulario-editar-cliente.html'
     instance = Cliente.objects.get(pk=pk)
-    form = ClienteForm(request.POST or None, instance=instance)
+    form = ClienteForm(request.POST or None, instance=instance, initial={'usuario':request.user})
+    
+    if instance.usuario != request.user:
+        raise PermissionError
 
     if request.method == 'POST':
         if form.is_valid():
@@ -43,14 +56,21 @@ def editarcliente(request, pk):
     return render(request, template_name, context)
 
 
+
+
+@login_required
 def apagarcliente(request, pk):
     template_name = 'cliente/tabela/tabela-cliente.html'
     obj = Cliente.objects.get(pk=pk)
-    obj.delete()
+    if obj.usuario != request.user:
+        raise PermissionError
+    else:
+        obj.delete()
     return render(request, template_name)
 
 
 
+@login_required
 def total_clientes(request):
     template_name = 'cliente/informacao-cliente.html'
     total_cliente =  Cliente.objects.count()
